Main code reorganization. I have splitted the code in different modules to improve maintainability and comprehension.
