+++ opencog/haskell/OpenCog/AtomSpace.hs
@@ -1,0 +1,19 @@
+
+module OpenCog.AtomSpace
+    ( AtomSpace
+    , runOnNewAtomSpace
+    , insert
+    , remove
+    , get
+    , debug
+    , TruthVal (..)
+    , AtomName (..)
+    , Atom (..)
+    , AtomGen (..)
+    , appAtomGen
+    ) where
+
+import OpenCog.AtomSpace.Api
+import OpenCog.AtomSpace.Types
+import OpenCog.AtomSpace.Env    (AtomSpace,runOnNewAtomSpace)
+
